pyproject.toml
[project]
name = "backend"
description = "This project powers the MaiVietLand backend."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Glinteco Developers", email = "developers@glinteco.com"}
]
version = "0.1.0"
dynamic = ["dependencies"]

[tool.ruff]
line-length = 119
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B904",  # Within an except clause, raise exceptions with raise ... from err
    "B007",  # Loop control variable not used within loop body
    "B017",  # assertRaises(Exception) should be considered evil
    "F841",  # Local variable is assigned to but never used
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["apps", "libs", "settings"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
ENVIRONMENT = "test"
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]
addopts = ["--tb=short", "-v", "--reuse-db"]

[tool.bandit]
exclude_dirs = [
    "**/tests/**",
    "**/test_*.py",
    "**/migrations/**",
    ".venv",
    "venv",
    "env",
    ".env",
    "**/migrations/**",
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "5.2.6"
django-extensions = "^3.2.3"
django-filter = "^25.1"
dj-database-url = "^2.3.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
easy-thumbnails = "^2.10"
sentry-sdk = "^2.24.1"

djangorestframework = "^3.15.2"
drf-standardized-errors = "^0.14.1"
djangorestframework-simplejwt = "^5.5.0"
drf-spectacular = "^0.28.0"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"

django-storages = {extras = ["s3"], version = "^1.14.5"}

celery = "^5.4.0"
flower = "^2.0.1"
django-celery-beat = "^2.7.0"
django-celery-results = "^2.5.1"
django-health-check = "^3.20.0"
django-waffle = "^5.0.0"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3"
pytest-django = "^4.10.0"
pytest-cov = "^6.0.0"
pytest-xdist = "^3.6.1"
ipython = "^9.0.2"
pre-commit = "^4.2.0"
factory-boy = "^3.3.3"
mypy = "^1.15.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.3"}
bandit = "^1.8.6"
ruff = "^0.13.2"
